// src/features/integration/resources/FormValues.ts
export type FormValues = {
    name: string,
    description: string,
    selectedForm: string;
    creationStrategy: string;
    caseId: string;
    title: string;
    publicTitle: string;
    caseType: string;
    administrativeUnit: string;
    archiveUnit: string;
    recordUnit: string;
    accessCode: string;
    paragraph: string;
    caseWorker: string;
    classification: string;
    primaryClass: string;
    secondaryClass: string;
    createdBy: string;

    recordTitle: string;
    recordPublictitle: string;
    recordCategory: string;
    recordAdministrativeUnit: string;
    recordStatus: string;
    recordCaseWorker: string;
    recordParagraph: string;

    documentTitle: string,
    documentAccessCode: string,
    documentParagraph: string,
    documentVariant: string,
    documentFormat: string,
    applicantName: string,
    applicantAddress: string;
    applicantPostalCode: string;
    applicantCity: string;
    applicantPhoneNumber: string;
    applicantEmail: string;
};